fixed server dependency problem
TLDR; pygame was required in common due to a type hint; causing problems

The server shouldn't require pygame to run only the client. Due to a
method on the `Vec2D` class using `pygame.rect` as a type hint, the
common package now required pygame which broke the tests and made the
server dependent on pygame.